Player data from poll

+++ src/main/java/com/cnc/model/resources/Resource.java
@@ -15,8 +15,8 @@ public class Resource {
         d = ((Number) data.get("d")).doubleValue();
     }
 
-    public double getValue() {
+    public int getValue() {
         long diffSteps = Client.getStep() - s;
-        return diffSteps * d + current;
+        return (int)(diffSteps * d + current);
     }
 }
